Handle every binary VDF field type and isolate bad app cache entries
The appinfo.vdf reader only understood BEGIN_OBJECT, STRING, and INT_32;
FLOAT_32, POINTER, WIDESTRING, COLOR, and INT_64 all threw, and since
the whole file was parsed as one sequential stream, any single entry
using one of those types would abort the entire read. Valve can and
does add fields over time, so this made the reader brittle against
formats we haven't seen yet rather than just the ones we have.

Each app entry already carries its own declared byte length, so
entries are now read into a bounded buffer and parsed from there. A
failure partway through one entry no longer disturbs the stream
position for the next, so an unparseable entry is logged and skipped
instead of losing every app in the cache. Reading entries this way
also exposed a real bug: parse64Int read 8 bytes but decoded only the
first 4 via ByteBuffer.getInt(), silently truncating the string-table
offset and picsToken. Both now go through a correct 64-bit reader, and
App.picsToken is a long. The bare asserts guarding BEGIN_OBJECT,
END_OBJECT, and the version marker were also no-ops without -ea;
they're now real exceptions.

AppCacheReader's single-thread-executor-plus-4s-timeout wrapper is
gone too. It was a stopgap from 2022 for a genuine while(true)
deadlock that was fixed at the loop level the same year; the app cache
path is always local, not cloud-synced, and with per-entry failures
now contained there's nothing left for the timeout to guard against.

+++ steam-config-reader/src/main/java/com/selesse/steam/appcache/App.java
@@ -5,7 +5,7 @@ public record App(
         int size,
         int infoState,
         int lastUpdated,
-        int picsToken,
+        long picsToken,
         byte[] sha1,
         int changeNumber,
         byte[] sha1Binary,
